+++ module/strategy/KickToGoal/src/KickToGoal.cpp
@@ -33,11 +33,11 @@ namespace module::strategy {
            Every<30, Per<std::chrono::seconds>>>()
             .then([this](const Field& field, const Sensors& sensors, const FieldDescription& field_description) {
                 // Get the robot's position (pose) on the field
-                Eigen::Isometry3d Hrf = Eigen::Isometry3d(sensors.Hrw) * Eigen::Isometry3d(field.Hfw.inverse());
+                Eigen::Isometry3f Hrf = Eigen::Isometry3f(sensors.Hrw) * Eigen::Isometry3f(field.Hfw.inverse());
 
                 // Get the goal position relative to the robot to kick to
-                Eigen::Vector3d rGFf = Eigen::Vector3d(-field_description.dimensions.field_length / 2.0, 0.0, 0.0);
-                Eigen::Vector3d rGRr = Hrf * rGFf;
+                Eigen::Vector3f rGFf = Eigen::Vector3f(-field_description.dimensions.field_length / 2.0, 0.0, 0.0);
+                Eigen::Vector3f rGRr = Hrf * rGFf;
 
                 emit<Task>(std::make_unique<KickTo>(rGRr));  // kick the ball if possible
             });
